Fix Netbricks push_header() bug
For some funny reason, the push_header() method implemented for the
Netbricks Packet<T, M> struct overwrites the supplied header with a
pointer to the beginning of the packet's payload - the very position
at which the header needs to be written. As a result, this method
allocates space for the new header, but does not write anything
because the source and destination on the write are the same!

Additionally, the push_header() method also "rewraps" the packet with
a new type signature after performing the write. The rewrapping function
takes a pointer to the inner-most network header. This is incorrectly set
to the passed in header rather than the position within the packet at
which the header was written to!