Implemented simple client for Sandstorm, not fully tested yet
This client has been tested only with tcpdump so far, and not
over an actual network.

This client is single threaded, and generates packets with a MAC,
IP, UDP, and RPC header. The RPC header corresponds to that of a
simple get() request. For now, all packets correspond to the same
30 B key.