Add Version metadata to every Table entry.
Each key has an Entry that holds a Version and a Bytes (value). The version
number is updated each time the value associated with a key is changed. If a
value for a key is deleted and then a new one is later inserted, the code
ensures that the newly inserted version is guaranteed to have a Version greater
than that of the Version that was removed.

By using these versions, code that works with the table can tell if values have
changed, which is useful for concurrency control.