Add scheduler for task execution on the client side.
The sequence is Send-Recv-Execute. When the client recieves an extension
with the pushback flag, it creates a new task and add it to queue. The tasks
are executed in round-robin fashion. Whenever, a task waits for IO, it is
removed from the queue by marking not ready and put in a hashmap. And when
the client recieves the response for any waiting tasks, it is again moved
backed to the round-robin queue.